* Implement MatrixDomain::trmm, in-place triangular matrix-multiply * New tests in test-matrix-domain.h for trmm * Add enum to indicate on which side a triangular matrix is * Add arguments start_idx and end_idx to internal gemvCol functions, allowing gemv to select only part of the input (not implemented for gemvRow) * Update GaussJordan to use trmm in GaussTransform, so it now requires no external storage(!) * Add trmm-specialisations in MatrixDomainM4RI, disabled for now * New struct RealMatrixType, generalising SubmatrixType (which is now gone) * Add copy-constructors for BitVectorReference and BitVectorConstReference * Change difference_type to ptrdiff_t in some places * Avoid bitwise operations between bitvector-references and bools, instead using if-statements * Fix parameters to test-matrix-domain and test-matrix-domain-gf2 so that first matrix is square by default

+++ linbox/vector/sparse.h
@@ -108,7 +108,7 @@ public:
 	typedef const SparseVectorReference<ConstIndexIterator, ConstElementIterator, ConstIndexIterator, ConstElementIterator> const_reference;
 	typedef reference *pointer;
 	typedef const_reference *const_pointer;
-	typedef long difference_type;
+	typedef ptrdiff_t difference_type;
 	typedef size_t size_type;
 
 	SparseVectorIterator () {}
